fix: address cdp trove review feedback
